DeepStore instances should (potentially) know about a name.
Scenario:
When DeepStore "A" has to instantiate another DeepStore "B", it does so
under the pretense that "B" is named. This is because of the tree
structure that occurs from doing such an operation. "A" knows the name
of "B" because it has to keep track of it, but "B" does not know its
own name.

This commit fixes the problem in the above scenario. This isn't
necessarily useful to DeepStore instances, but DeepStore is meant
to be subclassed.